requirements.txt
flask==1.1.2
flask-jwt-extended==3.24.1
psycopg2-binary==2.8.6
bcrypt==3.2.0
Werkzeug==1.0.1
uuid==1.30
moviepy==1.0.3
opencv-python==4.5.1.48
numpy==1.20.1
gunicorn==20.0.4